feat: move Actions into own component
