+++ app/app/game/styles/actions.module.scss
@@ -1,0 +1,48 @@
+.gameUI {
+  display: grid;
+  padding: 3rem;
+  gap: 3rem;
+  grid-template-columns: 30% 30% 30%;
+  justify-content: center;
+  height: 20vh;
+
+  & button {
+      background-color: grey;
+      color: white;
+      width: 15rem;
+      font-size: 1.1rem;
+      cursor: pointer;
+  }
+}
+
+.availableFunctions {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+  margin-left: 1rem;
+  margin-right: 1rem;
+  border: 2px solid black;
+}
+
+
+.chosenFunctions {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+  margin-left: 1rem;
+  margin-right: 1rem;
+  border: 2px dashed black;
+}
+
+.actions {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-left: 1rem;
+  margin-right: 1rem;
+  border: 2px solid black;
+}
