+++ app/mocks/inputSignals.ts
@@ -1,0 +1,40 @@
+export const mockInputSignals = {
+  initialGrid: [
+    [1, 1, 2, 0, 0, 0, 0, 0],
+    [1, 2, 0, 0, 0, 0, 0, 0],
+    [2, 0, 0, 0, 0, 0, 0, 0],
+    [2, 0, 0, 0, 0, 0, 0, 0],
+    [1, 2, 0, 0, 0, 0, 0, 0],
+    [1, 1, 2, 0, 0, 0, 0, 0],
+    [0, 0, 0, 0, 0, 0, 0, 0],
+    [0, 0, 0, 0, 0, 0, 0, 0],
+  ],
+  finalGrid: [
+    [1, 1, 1, 2, 0, 0, 0, 0],
+    [1, 1, 2, 0, 0, 0, 0, 0],
+    [1, 2, 0, 0, 0, 0, 0, 0],
+    [1, 2, 0, 0, 0, 0, 0, 0],
+    [1, 1, 2, 0, 0, 0, 0, 0],
+    [1, 1, 1, 2, 0, 0, 0, 0],
+    [0, 0, 0, 0, 0, 0, 0, 0],
+    [0, 0, 0, 0, 0, 0, 0, 0],
+  ],
+  account: "0x123",
+  selectedFunctions: [
+    [
+      [1, 2, 1],
+      [0, 0, 0],
+      [0, 0, 0],
+    ],
+    [
+      [0, 0, 0],
+      [1, 2, 0],
+      [0, 0, 0],
+    ],
+    [
+      [0, 0, 0],
+      [0, 0, 0],
+      [0, 0, 0],
+    ],
+  ],
+}
